Parallel: implement modules of init, slot executer and dispatcher
BEP 130 parallel transaction execution will maintain a tx execution routine
pool, a configured number of slot(routine) to execution transactions.

Init is executed once on startup and will create the routine pool.
Slot executer is the place to execute transactions.
The dispacther is the module that will dispatch transaction to the right slot.