Parallel BugFix: fix a fork on block 15,338,563 (Part 2)
Previous commit didn't fix the fork.
It a nonce conflict caused by opcode: opCreate & opCreate2
Generally, the nonce is advanced by 1 for the transaction sender;
But opCreate & opCreate2 will try to create a new contract, the
caller will advance its nonce too.

It makes the nonce conflict detect more complicated: as nonce is a
fundamental part of an account, as long as it has been changed, we mark
the address as StateChanged, any concurrent access to it will be considered
as conflicted.