Parallel: optimize conflict for AddBalance(0)
Add balance with 0 did nothing, but it will do an empty() check, and add
a touch event. Add on transaction finalize, the touch event will check if
the StateObject is empty, do empty delete if it is.

This patch is to take the empty check as a state check, if the addr state has
not been changed(create, suicide, empty delete), then empty check is reliable.