pbs: fix a inaccurate bid result log
sometimes "bestBid == nil" does not mean it is the first bid,
if new block was just importing during the bid simulation, the bestBid
will be clear as well. In this case, the bid is stale and useless.

Keep the SetBestBid right now, as it will be cleared later anyway.
Once we have more confidence, it can be removed