pbs: fix a inaccurate bid result log (#2959)
sometimes "bestBid == nil" does not mean it is the first bid,
if new block was just imported during the bid simulation, the bestBid
will be cleared as well. In this case, the bid is stale and useless.

Keep the SetBestBid right now, as it will be cleared later anyway.
Once we have more confidence, it can be removed